+++ lib/src/generators/guilds/mottos/guild_motto_generator.dart
@@ -1,3 +1,4 @@
+import '../../../strings_manipulations.dart';
 import '../../base/batch_generator.dart';
 import '../../base/generator.dart';
 import '../../base/list_item_generator.dart';
@@ -34,10 +35,10 @@ class GuildMottoGenerator implements IGenerator<String> {
       "through ${map["traits"].join(" and ")}, we ${map["action"]} to ${map["goal"]}";
 
   String _getAlternativeMotto(Map<String, dynamic> map) =>
-      "${map["action"]}ing with ${map["traits"].join(" and ")} for ${map["goal"].split(" ").first}ing ${map["goal"].split(" ").sublist(1).join(" ")}";
+      "${toGerund(map["action"])} with ${map["traits"].join(" and ")} for ${toGerund(map["goal"].split(" ").first)} ${map["goal"].split(" ").sublist(1).join(" ")}";
 
   String _getAbstractMotto(Map<String, dynamic> map) =>
-      "${map["goal"].split(" ").first}ing ${map["goal"].split(" ").sublist(1).join(" ")} with ${map["traits"].join(" and ")}";
+      "${toGerund(map["goal"].split(" ").first)} ${map["goal"].split(" ").sublist(1).join(" ")} with ${map["traits"].join(" and ")}";
 
   Map<String, IGenerator> _getBatch() => {
         "traits": UniqueGenerator(