+++ lib/src/subtypes/kingdoms/kingdom_type.dart
@@ -2,7 +2,6 @@ import '../../generators/base/generator.dart';
 import '../guilds/guild_type.dart';
 import '../races/race.dart';
 import '../settlements/settlement_type.dart';
-import 'government_types/government_type.dart';
 
 /// A class that represents a generic kingdom type
 abstract class KingdomType {
@@ -20,9 +19,6 @@ abstract class KingdomType {
 
   /// Returns a generator for the type of other important settlements in the kingdom
   IGenerator<SettlementType> getImportantSettlementsTypesGenerator();
-
-  /// Returns a generator for the government type of the kingdom
-  IGenerator<GovernmentType> getGovernmentTypeGenerator();
 
   /// Returns a generator for something a kingdom is known for
   IGenerator<String> getKnownForGenerator();